Fixed about screen and added OverlayEventListener.
